add view padding

+++ arcgis_map_android/android/src/main/kotlin/esri/arcgis/flutter_plugin/model/ViewPadding.kt
@@ -1,8 +1,8 @@
 package esri.arcgis.flutter_plugin.model
 
 class ViewPadding(
-    val left: Double,
-    val top: Double,
-    val right: Double,
-    val bottom: Double,
+    val left: Int,
+    val top: Int,
+    val right: Int,
+    val bottom: Int,
 )
